python-object_relational_mapping/requirements.txt
Flask==2.0.2
SQLAlchemy==2.0.30
mysqlclient==2.1.1
